+++ Samples/UniFlux.Sample.5/Sample_5.cs
@@ -43,7 +43,7 @@ namespace UniFlux.Sample
             color_current = color;
             history_colors.Add(color);
         }
-        [Flux(nameof(Sample_5) + ".ChangePrimary_Color1")] private void _ChangePrimary_Color1() => K_Primary.DispatchState(color_1);
-        [Flux(nameof(Sample_5) + ".ChangePrimary_Color2")] private void _ChangePrimary_Color2() => K_Primary.DispatchState(color_2);
+        [MethodFlux(nameof(Sample_5) + ".ChangePrimary_Color1")] private void _ChangePrimary_Color1() => K_Primary.DispatchState(color_1);
+        [MethodFlux(nameof(Sample_5) + ".ChangePrimary_Color2")] private void _ChangePrimary_Color2() => K_Primary.DispatchState(color_2);
     }
 }